feat: Add Remote Storage module documentation
The "remotestorage.mdx" file was modified to add documentation for the Remote Storage module. This change includes information on how to get stored files, read a file, and access extra functions related to remote storage. The documentation provides users with a clear understanding of how to use the Remote Storage module in their projects.